@@ -29,7 +29,7 @@
     "test-cov": "nyc --reporter=text mocha test/spec/**/*.js"
   },
   "main": "./index.js",
-  "version": "6.0.90",
+  "version": "7.0.0",
   "license": "MIT",
   "repository": {
     "type": "git",
@@ -38,32 +38,28 @@
   "readme": "",
   "readmeFilename": "README.md",
   "dependencies": {
-    "appolo-agent": "^6.0.37",
-    "appolo-context": "^6.0.9",
-    "appolo-engine": "^6.0.79",
-    "bluebird": "^3.5.5",
-    "joi": "^14.3.1",
-    "lodash": "^4.17.15",
-    "tslib": "^1.10.0"
+    "appolo-agent": "^7.0.0",
+    "appolo-engine": "^7.0.0",
+    "appolo-utils": "^0.0.47",
+    "tslib": "^1.11.1"
   },
   "devDependencies": {
     "@types/benchmark": "^1.0.31",
-    "@types/bluebird": "^3.5.27",
-    "@types/chai": "^4.1.7",
-    "@types/chai-http": "4.2.0",
+    "@types/bluebird": "^3.5.30",
+    "@types/chai": "^4.2.11",
+    "@types/chai-http": "^4.2.0",
     "@types/cookie": "^0.3.3",
-    "@types/cookie-parser": "^1.4.1",
-    "@types/joi": "^14.3.3",
-    "@types/lodash": "^4.14.135",
-    "@types/mocha": "^5.2.7",
-    "@types/node": "^12.0.10",
-    "@types/qs": "^6.5.3",
-    "@types/serve-static": "^1.13.2",
-    "@types/sinon": "^7.0.13",
-    "@types/sinon-chai": "^3.2.2",
-    "@types/supertest": "^2.0.7",
-    "@types/type-is": "^1.6.2",
-    "autocannon": "^4.0.0",
+    "@types/cookie-parser": "^1.4.2",
+    "@types/lodash": "^4.14.149",
+    "@types/mocha": "^7.0.2",
+    "@types/node": "^13.9.1",
+    "@types/qs": "^6.9.1",
+    "@types/serve-static": "^1.13.3",
+    "@types/sinon": "^7.5.2",
+    "@types/sinon-chai": "^3.2.3",
+    "@types/supertest": "^2.0.8",
+    "@types/type-is": "^1.6.3",
+    "autocannon": "^4.6.0",
     "benchmark": "^2.1.4",
     "body-parser": "^1.19.0",
     "chai": "4.2.0",
@@ -71,16 +67,16 @@
     "cli-table": "^0.3.1",
     "compression": "^1.7.4",
     "consolidate": "^0.15.1",
-    "cookie-parser": "^1.4.4",
+    "cookie-parser": "^1.4.5",
     "express": "^4.17.1",
-    "fastify": "^2.6.0",
-    "mocha": "^6.1.4",
-    "nunjucks": "^3.2.0",
+    "fastify": "^2.12.1",
+    "mocha": "^7.1.1",
+    "nunjucks": "^3.2.1",
     "nyc": "^15.0.0",
     "serve-static": "^1.14.1",
-    "sinon": "^7.3.2",
-    "sinon-chai": "^3.3.0",
+    "sinon": "^9.0.1",
+    "sinon-chai": "^3.5.0",
     "supertest": "^4.0.2",
-    "typescript": "^3.5.2"
+    "typescript": "^3.8.3"
   }
 }
