@@ -1,5 +1,5 @@
 import appolo = require('../../../../index');
-import Q = require('bluebird');
+import {Promises} from 'appolo-utils';
 
 
 export function logger3(options?) {
@@ -9,11 +9,11 @@ export function logger3(options?) {
             getName: function () {
                 return env.test + "logger3";
             }
-        }
+        };
 
         inject.addObject('logger3', logger3);
 
-        return Q.delay(1);
+        return Promises.delay(1);
     }
 }
 
