@@ -2,10 +2,8 @@
 Object.defineProperty(exports, "__esModule", { value: true });
 const appolo_agent_1 = require("appolo-agent");
 const appolo_engine_1 = require("appolo-engine");
-const appolo_context_1 = require("appolo-context");
 const launcher_1 = require("./launcher/launcher");
 const route_1 = require("./routes/route");
-const IMiddleware_1 = require("./interfaces/IMiddleware");
 const util_1 = require("./util/util");
 const invokeActionMiddleware_1 = require("./routes/invokeActionMiddleware");
 class App extends appolo_engine_1.EventDispatcher {
@@ -41,15 +39,19 @@ class App extends appolo_engine_1.EventDispatcher {
     get options() {
         return this._launcher.options;
     }
-    enableContext(contextCtr) {
-        let context = appolo_context_1.namespace.create(IMiddleware_1.RequestContextSymbol, contextCtr);
-        this.injector.addObject("context", context);
-        context.initialize();
-        this.use((req, res, next) => context.scope(next));
-    }
-    getContext() {
-        return appolo_context_1.namespace.get(IMiddleware_1.RequestContextSymbol);
-    }
+    // public enableContext(contextCtr?: typeof Context) {
+    //
+    //     let context = namespace.create(RequestContextSymbol, contextCtr);
+    //
+    //     this.injector.addObject("context", context);
+    //
+    //     context.initialize();
+    //
+    //     this.use((req: IRequest, res: IResponse, next: NextFn) => context.scope(next))
+    // }
+    // public getContext() {
+    //     return namespace.get(RequestContextSymbol);
+    // }
     use(path, ...middleware) {
         return this._addMiddleware(path, middleware, false);
     }
@@ -97,7 +99,6 @@ class App extends appolo_engine_1.EventDispatcher {
         this._launcher.options[name] = value;
     }
     async reset() {
-        appolo_context_1.namespace.delete(IMiddleware_1.RequestContextSymbol);
         await this._launcher.reset();
     }
     register(id, type) {
