@@ -32,7 +32,7 @@ class StaticController {
         throw new appolo_agent_1.NotFoundError(error, data, code);
     }
     getModel(req) {
-        return req.model;
+        return Object.assign({}, req.body || {}, req.query || {}, req.params || {});
     }
 }
 exports.StaticController = StaticController;
