@@ -1,10 +1,9 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-const _ = require("lodash");
 const route_1 = require("../routes/route");
 const appolo_agent_1 = require("appolo-agent");
 const appolo_engine_1 = require("appolo-engine");
-const routeModel_1 = require("../routes/routeModel");
+const appolo_utils_1 = require("appolo-utils");
 const util_1 = require("../util/util");
 exports.RouterDefinitionsSymbol = "__RouterDefinitions__";
 exports.RouterDefinitionsCompiledSymbol = "__RouterDefinitionsCompiled__";
@@ -13,7 +12,7 @@ exports.RouterModelSymbol = "__RouterModelDefinitions__";
 exports.RouterControllerSymbol = "__RouterControllerDefinitions__";
 function defineRouteClass(params, target) {
     let route = util_1.Util.getReflectData(exports.RouterDefinitionsClassSymbol, target, new route_1.Route(target));
-    _.forEach(params, param => {
+    (params || []).forEach(param => {
         route[param.name].apply(route, param.args);
     });
 }
@@ -35,7 +34,7 @@ function defineRouteProperty(params) {
             data[propertyKey] = route = new route_1.Route(target.constructor);
             route.action(propertyKey);
         }
-        _.forEach(params, param => {
+        (params || []).forEach(param => {
             route[param.name].apply(route, param.args);
         });
     }.bind(null, params);
@@ -78,33 +77,19 @@ function hook(name, ...hook) {
 }
 exports.hook = hook;
 function middleware(middleware) {
-    if (_.isArray(middleware)) {
-        middleware = _(middleware).clone().reverse();
+    if (Array.isArray(middleware)) {
+        middleware = appolo_utils_1.Arrays.clone(middleware).reverse();
     }
     return defineRouteProperty([{ name: "middleware", args: [middleware, "head"] }]);
 }
 exports.middleware = middleware;
 function error(middleware) {
-    if (_.isArray(middleware)) {
-        middleware = _(middleware).clone().reverse();
+    if (Array.isArray(middleware)) {
+        middleware = appolo_utils_1.Arrays.clone(middleware).reverse();
     }
     return defineRouteProperty([{ name: "error", args: [middleware, "head"] }]);
 }
 exports.error = error;
-function validation(key, validation) {
-    if (key.constructor && key.constructor.prototype === routeModel_1.RouteModel.constructor.prototype && key.prototype && Reflect.hasMetadata(exports.RouterModelSymbol, key)) {
-        key = Reflect.getMetadata(exports.RouterModelSymbol, key);
-    }
-    return defineRouteProperty([{ name: "validation", args: [key, validation] }]);
-}
-exports.validation = validation;
-function validationParam(validation) {
-    return function (target, propertyKey, descriptor) {
-        let validations = util_1.Util.getReflectData(exports.RouterModelSymbol, target.constructor, {});
-        validations[propertyKey] = validation;
-    };
-}
-exports.validationParam = validationParam;
 function abstract(route) {
     util_1.Util.reverseMiddleware(route);
     return defineRouteProperty([{ name: "abstract", args: [route] }]);
