@@ -4,7 +4,7 @@ const appolo_engine_1 = require("appolo-engine");
 const appolo_agent_1 = require("appolo-agent");
 const router_1 = require("../routes/router");
 const path = require("path");
-const _ = require("lodash");
+const appolo_utils_1 = require("appolo-utils");
 const decorators_1 = require("../decorators/decorators");
 const util_1 = require("../util/util");
 const decorate_1 = require("../routes/decorate");
@@ -18,18 +18,13 @@ let Defaults = {
     urlParser: "fast",
     viewExt: "html",
     viewEngine: null,
-    viewFolder: "src/views",
-    validatorOptions: {
-        abortEarly: false,
-        allowUnknown: true,
-        stripUnknown: true
-    }
+    viewFolder: "src/views"
 };
 class Launcher {
     constructor(options, _app) {
         this._app = _app;
         this._plugins = [];
-        this._options = _.defaults({}, options, Defaults);
+        this._options = appolo_utils_1.Objects.defaults({}, options || {}, Defaults);
         this._engine = new appolo_engine_1.App(this._options);
         this._agent = new appolo_agent_1.Agent(this._options);
         this._port = this._getPort();
@@ -54,7 +49,6 @@ class Launcher {
         this._app.on(appolo_engine_1.Events.ModuleExport, this.addRoute, this);
         await this._engine.launch();
         await this.loadCustomConfigurations();
-        //_.forEach(this._app.exported,exported=>this.addRoute(exported.fn));
         this._router.initialize();
         await this._agent.listen(this._port);
         this._logServerMessage();
@@ -68,18 +62,20 @@ class Launcher {
         if (!routeData || !Reflect.hasOwnMetadata(decorators_1.RouterControllerSymbol, fn)) {
             return;
         }
-        routeData = _.cloneDeep(routeData);
-        routeAbstract = _.cloneDeep(routeAbstract);
-        _.forEach(routeData, (route, key) => {
+        routeData = appolo_utils_1.Objects.cloneDeep(routeData);
+        Object.keys(routeData).forEach((key) => {
+            let route = routeData[key];
+            route = route.clone();
             //add abstract route
             if (routeAbstract) {
+                routeAbstract = routeAbstract.clone();
                 util_1.Util.reverseMiddleware(routeAbstract.definition);
                 route.abstract(routeAbstract.definition);
             }
             let prefix = Reflect.getOwnMetadata(decorators_1.RouterControllerSymbol, fn);
             //add prefix to routes
             if (prefix) {
-                _.forEach(route.definition.path, (_path, index) => {
+                (route.definition.path || []).forEach((_path, index) => {
                     route.definition.path[index] = path.join("/", prefix, _path);
                 });
             }
@@ -97,7 +93,7 @@ class Launcher {
         await util_1.Util.loadPathWithArgs([allPath, environmentPath], this._engine.injector);
     }
     _logServerMessage() {
-        let msg = _.template(this._options.startMessage, { interpolate: /\${([\s\S]+?)}/g })({
+        let msg = appolo_utils_1.Strings.replaceFormat(this._options.startMessage, {
             port: this._port,
             version: this._engine.env.version,
             environment: this._engine.env.type
