@@ -1,6 +1,6 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-const _ = require("lodash");
+const appolo_utils_1 = require("appolo-utils");
 const _path = require("path");
 function decorate(req, res, app) {
     let old = res.render;
@@ -12,8 +12,8 @@ function decorate(req, res, app) {
         if (!path && this.req.route) {
             path = _path.resolve(this.req.app.options.root, "src/controllers", this.req.route.controllerName, this.req.route.actionName);
         }
-        let paths = _.isArray(path) ? path : [path];
-        if (_.isString(path) && this.req.route) {
+        let paths = Array.isArray(path) ? path : [path];
+        if (appolo_utils_1.Strings.isString(path) && this.req.route) {
             paths.push(_path.resolve(this.req.app.options.root, "src/controllers", this.req.route.controllerName, path));
         }
         return old.call(this, paths, params);
