@@ -1,7 +1,7 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
 const index_1 = require("appolo-agent/index");
-const _ = require("lodash");
+const appolo_utils_1 = require("appolo-utils");
 function invokeActionMiddleware(req, res, next) {
     let route = req.route;
     let controller = req.app.injector.getObject(route.controller, [req, res, route]);
@@ -11,7 +11,7 @@ function invokeActionMiddleware(req, res, next) {
     }
     let fnName = route.actionName;
     if (!fnName) {
-        fnName = _.isString(route.action) ? route.action : route.action(controller).name;
+        fnName = appolo_utils_1.Strings.isString(route.action) ? route.action : route.action(controller).name;
         if (!controller[fnName]) {
             next(new index_1.HttpError(500, `failed to invoke ${this.constructor.name} fnName ${fnName}`));
             return;
