@@ -1,18 +1,17 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-const _ = require("lodash");
+const appolo_utils_1 = require("appolo-utils");
 const util_1 = require("../util/util");
-const routeModel_1 = require("./routeModel");
 let orderIndex = 0;
 class Route {
     constructor(controller) {
+        this._controller = controller;
         this._route = {
             method: [],
             roles: [],
             environments: [],
             middleware: [],
             middlewareError: [],
-            validations: {},
             controller: util_1.Util.getControllerName(controller),
             path: [],
             order: orderIndex++,
@@ -56,14 +55,14 @@ class Route {
         return this;
     }
     abstract(abstract) {
-        let items = _.pick(abstract, ["environments", "roles", "middleware", "validations", "convertToCamelCase", "params"]);
-        _.forEach(items, (item, key) => {
-            this[key](item);
+        let items = appolo_utils_1.Objects.pick(abstract, "environments", "roles", "middleware", "convertToCamelCase", "params");
+        Object.keys(items).forEach(key => {
+            this[key](items[key]);
         });
         return this;
     }
     extend(opts) {
-        _.extend(this._route, opts);
+        Object.assign(this._route, opts);
         return this;
     }
     param(key, value) {
@@ -71,24 +70,29 @@ class Route {
         return this;
     }
     params(params) {
-        _.defaults(this._route.params, params);
+        appolo_utils_1.Objects.defaults(this._route.params, params || {});
         return this;
     }
-    validation(key, validation) {
-        return this.validations(key, validation);
-    }
-    validations(key, validation) {
-        if (key.constructor && key.constructor.prototype === routeModel_1.RouteModel.constructor.prototype && key.prototype && key.prototype.__validations__) {
-            key = key.prototype.__validations__;
-        }
-        if (_.isObject(key)) {
-            _.extend(this._route.validations, key);
-        }
-        else {
-            this._route.validations[key] = validation;
-        }
-        return this;
-    }
+    // public validation(key: string | { [index: string]: joi.Schema } | RouteModel, validation?: joi.Schema): this {
+    //     return this.validations(key, validation);
+    // }
+    // public validations(key: string | { [index: string]: joi.Schema } | RouteModel, validation?: joi.Schema): this {
+    //
+    //     if (key.constructor && key.constructor.prototype === RouteModel.constructor.prototype && (key as any).prototype && (key as any).prototype.__validations__) {
+    //         key = (key as any).prototype.__validations__
+    //     }
+    //
+    //     if (_.isObject(key)) {
+    //
+    //         _.extend(this._route.validations, key)
+    //
+    //     } else {
+    //
+    //         this._route.validations[key as string] = validation
+    //     }
+    //
+    //     return this;
+    // }
     method(method) {
         this._route.method.push(method);
         return this;
@@ -97,7 +101,7 @@ class Route {
         return this.environments(environment);
     }
     environments(environment) {
-        if (_.isArray(environment)) {
+        if (Array.isArray(environment)) {
             this._route.environments.push.apply(this._route.environments, environment);
         }
         else {
@@ -118,11 +122,11 @@ class Route {
     _addMiddleware(middleware, order = "tail", error = false) {
         let arrMethod = order == "head" ? "unshift" : "push";
         //
-        if (_.isArray(middleware)) {
+        if (Array.isArray(middleware)) {
             return this.middlewares(middleware, order);
         }
         let middle = middleware;
-        if (_.isPlainObject(middle) && (middle.order && middle.middleware)) {
+        if (appolo_utils_1.Objects.isPlain(middle) && (middle.order && middle.middleware)) {
             return this.middleware(middle.middleware, middle.order);
         }
         let id = util_1.Util.getClassId(middle);
@@ -133,7 +137,7 @@ class Route {
         return this;
     }
     middlewares(middlewares, order = "tail") {
-        _.forEach(_.isArray(middlewares) ? middlewares : [middlewares], fn => this.middleware(fn, order));
+        appolo_utils_1.Arrays.arrayify(middlewares).forEach(fn => this.middleware(fn, order));
         return this;
     }
     addHook(name, ...hook) {
@@ -144,7 +148,7 @@ class Route {
         return this.roles(role);
     }
     roles(role) {
-        if (_.isArray(role)) {
+        if (Array.isArray(role)) {
             this._route.roles.push.apply(this._route.roles, role);
         }
         else {
@@ -166,12 +170,17 @@ class Route {
     }
     customRouteParam(index, fn) {
         this._route.customRouteParam.push({ index, fn });
-        this._route.customRouteParam = _.orderBy(this._route.customRouteParam, data => data.index);
+        this._route.customRouteParam = appolo_utils_1.Arrays.sortBy(this._route.customRouteParam, data => data.index);
         return this;
     }
     statusCode(code) {
         this._route.statusCode = code;
         return this;
+    }
+    clone() {
+        let route = new Route(this._controller);
+        route._route = appolo_utils_1.Objects.cloneDeep(this._route);
+        return route;
     }
 }
 exports.Route = Route;
