@@ -1,10 +1,8 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-const _ = require("lodash");
 const appolo_agent_1 = require("appolo-agent");
 const IMiddleware_1 = require("../interfaces/IMiddleware");
 const invokeActionMiddleware_1 = require("./invokeActionMiddleware");
-const checkValidationMiddleware_1 = require("./checkValidationMiddleware");
 const util_1 = require("../util/util");
 class Router {
     constructor(_env, _injector, _options, _agent) {
@@ -19,10 +17,10 @@ class Router {
     }
     initialize() {
         this._isInitialize = true;
-        _.forEach(this._routes, route => this._initRoute(route));
+        this._routes.forEach(route => this._initRoute(route));
     }
     getRoute(path, method) {
-        return _.find(this._routes, route => _.includes(route.definition.path, path) && _.includes(route.definition.method, method));
+        return this._routes.find(route => route.definition.path.includes(path) && route.definition.method.includes(method));
     }
     addRoute(route) {
         this._routes.push(route);
@@ -37,7 +35,7 @@ class Router {
         }
         def.$initialized = true;
         //check if we have valid path
-        if (!def.path.length || !def.action || (def.environments.length && !_.includes(def.environments, (this._env.name || this._env.type)))) {
+        if (!def.path.length || !def.action || (def.environments.length && def.environments.indexOf(this._env.name || this._env.type) == -1)) {
             return;
         }
         def.controllerName = def.controller.replace(this.controllerSuffix, '');
@@ -46,12 +44,15 @@ class Router {
         if (def.gzip || def.statusCode || def.headers.length || def.customRouteFn.length) {
             middewares.unshift(invokeActionMiddleware_1.invokeCustomRouteMiddleWare);
         }
-        if (!_.isEmpty(def.validations)) {
-            middewares.unshift(checkValidationMiddleware_1.checkValidationMiddleware);
-        }
+        // if (!_.isEmpty(def.validations)) {
+        //     middewares.unshift(checkValidationMiddleware);
+        // }
         middewares.push(invokeActionMiddleware_1.invokeActionMiddleware);
         let hooks = {};
-        _.forEach(def.hooks, (hook, key) => hooks[key] = util_1.Util.convertMiddlewareHooks(key, hook));
+        Object.keys(def.hooks || {}).forEach(key => {
+            let hook = def.hooks[key];
+            hooks[key] = util_1.Util.convertMiddlewareHooks(key, hook);
+        });
         for (let i = 0, len = def.path.length; i < len; i++) {
             this._agent.add(def.method[i] || appolo_agent_1.Methods.GET, def.path[i], middewares, def, hooks);
         }
