@@ -1,7 +1,7 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
 const appolo = require("appolo-engine");
-const _ = require("lodash");
+const appolo_utils_1 = require("appolo-utils");
 const appolo_agent_1 = require("appolo-agent");
 const IMiddleware_1 = require("../interfaces/IMiddleware");
 const invokeActionMiddleware_1 = require("../routes/invokeActionMiddleware");
@@ -25,18 +25,19 @@ class Util extends appolo.Util {
         return typeof v === 'function' && v.name && /^\s*class\s+/.test(v.toString());
     }
     static reverseMiddleware(route) {
-        _.forEach(route, (value, key) => {
+        Object.keys(route || {}).forEach(key => {
+            let value = route[key];
             //we need to insert middlewares in reverse order
             if (key == "middleware") {
                 route[key] = {
-                    middleware: _.isArray(value) ? value.reverse() : value,
+                    middleware: Array.isArray(value) ? value.reverse() : value,
                     order: "head"
                 };
             }
         });
     }
     static getControllerName(controller) {
-        return _.isFunction(controller) && controller.name ? _.camelCase(controller.name) : controller;
+        return appolo_utils_1.Functions.isFunction(controller) && controller.name ? Util.getClassName(controller) : controller;
     }
     static decorateRequest(name, fn) {
         appolo_agent_1.Request.prototype[name] = function () {
@@ -49,7 +50,7 @@ class Util extends appolo.Util {
         };
     }
     static getRouteDefinition(fn, action) {
-        action = _.isString(action) ? action : action(fn.prototype).name;
+        action = appolo_utils_1.Strings.isString(action) ? action : action(fn.prototype).name;
         let route = Reflect.getMetadata(decorators_1.RouterDefinitionsCompiledSymbol, fn, action);
         return route;
     }
