@@ -1,6 +1,6 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-const Q = require("bluebird");
+const appolo_utils_1 = require("appolo-utils");
 function logger2(options) {
     return async function (env, inject, logger) {
         let logger2 = {
@@ -9,7 +9,7 @@ function logger2(options) {
             }
         };
         inject.addObject('logger2', logger2);
-        return Q.delay(100);
+        return appolo_utils_1.Promises.delay(100);
     };
 }
 exports.logger2 = logger2;
