@@ -1,6 +1,6 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-const Q = require("bluebird");
+const appolo_utils_1 = require("appolo-utils");
 function logger3(options) {
     return async function (env, inject) {
         let logger3 = {
@@ -9,7 +9,7 @@ function logger3(options) {
             }
         };
         inject.addObject('logger3', logger3);
-        return Q.delay(1);
+        return appolo_utils_1.Promises.delay(1);
     };
 }
 exports.logger3 = logger3;
