@@ -1,9 +1,7 @@
 "use strict";
 import {inject, define, bootstrap, singleton, IBootstrap, override, Util} from "../../../index";
 import {Manager} from "./manager/manager";
-import * as Q from 'bluebird';
-import {HooksController} from "./controllers/hooksController";
-import {Hooks} from "appolo-agent/lib/types";
+import {Promises} from 'appolo-utils';
 
 @define()
 @bootstrap()
@@ -11,14 +9,14 @@ import {Hooks} from "appolo-agent/lib/types";
 @override()
 export class Bootstrap implements IBootstrap {
 
-    @inject() manager: Manager
+    @inject() manager: Manager;
 
     working: boolean;
 
     public async run() {
         this.working = true;
 
-        await Q.delay(10);
+        await Promises.delay(10);
 
     }
 }
