@@ -8,7 +8,7 @@ export class DefineController extends Controller {
     @inject() manager5: Manager5;
 
     test(req, res) {
-        res.json({working: true, controllerName: this.route.controller, model: req.model, manager5: this.manager5.name})
+        res.json({working: true, controllerName: this.route.controller, model: req.query, manager5: this.manager5.name})
     }
 }
 
