@@ -3,13 +3,13 @@ Object.defineProperty(exports, "__esModule", { value: true });
 const tslib_1 = require("tslib");
 const index_1 = require("../../../../index");
 const compression = require("compression");
-const Q = require("bluebird");
+const appolo_utils_1 = require("appolo-utils");
 let someHeader = index_1.customRouteDecorator((req, res, route) => {
     res.setHeader("x-test2", "222");
 });
 let GzipController = class GzipController extends index_1.Controller {
     async gzipAsync(req, res) {
-        await Q.delay(10);
+        await appolo_utils_1.Promises.delay(10);
         res.gzip().json({ working: true });
     }
     gzip(req, res) {
