@@ -7,8 +7,6 @@ import {
     post,
     singleton,
     StaticController,
-    validation,
-    validator
 } from '../../../../index';
 
 @controller()
@@ -18,23 +16,12 @@ export class StaticController2 extends StaticController {
     @inject() manager: any;
 
     @post("/test/static/controller/:name/:bbb/post")
-    @validation("test", validator.string())
-    @validation("name", validator.string())
-    @validation("userName", validator.string())
-    @validation("name2", validator.string())
-    @validation("name", validator.string())
-    @validation("testPost", validator.boolean().required())
-    test(req, res) {
+       test(req, res) {
         res.json({model: this.getModel(req)})
     }
 
     @get("/test/static/controller/:name/:name2")
-    @validation("test", validator.string())
-    @validation("name", validator.string())
-    @validation("userName", validator.string())
-    @validation("name2", validator.string())
-    @validation("name", validator.string())
-    test2(req, res) {
+        test2(req, res) {
         res.json({model: this.getModel(req)})
     }
 
