@@ -2,7 +2,7 @@
 import {define, singleton, inject, lazy,injectAlias} from '../../../../index' ;
 import {Manager3} from "./manager3";
 import {IHandler} from "../handlers/IHandler";
-import * as _ from 'lodash'
+import {Arrays} from "appolo-utils";
 
 
 @define()
@@ -14,7 +14,7 @@ export class Manager5 {
     @injectAlias("IHandler") handlers: IHandler[]
 
     public get name(): string {
-        return this.constructor.name + _.sumBy(this.handlers, h => h.handle()).toString()
+        return this.constructor.name + Arrays.sumBy(this.handlers, h => h.handle()).toString()
     }
 }
 
